🐛 fix: socket mocking 상태에서 사진을 보냈을 시, 사진 상세보기가 가능하도록 변경
- chatData를 공유하도록 함
