+++ src/mocks/chat/chatHandler.ts
@@ -1,10 +1,14 @@
 import { v4 as uuidv4 } from 'uuid';
+import Server from 'socket.io-mock';
+import SOCKET_KEY from 'constants/socketKeys';
+import { ImgDefaultProfile } from 'assets/image';
 import {
   NullableResponse,
   createApiHandler,
   getSearchParams,
 } from 'mocks/createApiHandler';
 import {
+  CreateChattingDto,
   ChatPhotoDto,
   ChatPhotoLineDto,
   ChatRequestDto,
@@ -23,7 +27,71 @@ import { originData } from './data/chatData';
 let chatData: ParentChildChattingDto[] = originData;
 
 const findChatData = (id: string): ChattingDto =>
-  originData.filter((chat) => chat.parentChatting.id === id)[0].parentChatting;
+  chatData.filter((chat) => chat.parentChatting.id === id)[0].parentChatting;
+
+// socket 관련
+class MockedSocket {
+  private server;
+
+  constructor() {
+    this.server = new Server();
+
+    this.server.on('connection', () => {
+      console.log('User connected');
+
+      this.server.on('disconnection', () => {
+        console.log('User disconnected');
+      });
+
+      this.server.on(SOCKET_KEY.CREATE_CHAT, (dto: CreateChattingDto) => {
+        console.log('get chat');
+
+        const data: ChattingDto = {
+          id: uuidv4(),
+          content: dto.content,
+          emojiUrl: dto.emojiId && dto.emojiId,
+          imageUrls: new Array(dto.imageIds.length).fill(ImgDefaultProfile),
+          videoUrls: [],
+          voiceMsgUrls: [],
+          createdAt: new Date(),
+          isMine: true,
+          Writer: {
+            id: '1',
+            name: '연주',
+            imageUrl: ImgDefaultProfile,
+          },
+        };
+        chatData = [...chatData, { parentChatting: data, childChatting: null }];
+
+        this.server.emit(SOCKET_KEY.GET_CHAT, data);
+      });
+    });
+
+    this.connect();
+  }
+
+  private connect() {
+    this.server.socketClient.emit('connection');
+  }
+
+  public disconnect() {
+    this.server.socketClient.emit('disconnection');
+  }
+
+  public on(event: string, callback: (res: any) => void) {
+    this.server.socketClient.on(event, callback);
+  }
+
+  public off(event: string) {
+    this.server.socketClient.off(event);
+  }
+
+  public emit(event: string, dto: any) {
+    this.server.socketClient.emit(event, dto);
+  }
+}
+
+export const socket = new MockedSocket();
 
 // chatting 관련
 
@@ -318,8 +386,10 @@ export const getPhotoDetailHandler = createApiHandler<
   path: '/couples/:coupleId/chattings/:chattingId/photos',
   method: 'get',
   requestHandler: ({ chattingId }) => {
+    console.log('111');
     const { id, createdAt, Writer, imageUrls, videoUrls } =
       findChatData(chattingId);
+    console.log('222');
 
     const data: ChatPhotoDto = {
       id,
